+++ source/scener/graphics/model_mesh_part.cpp
@@ -20,22 +20,22 @@ namespace scener::graphics
         return _effect.get();
     }
 
-    std::size_t model_mesh_part::start_index() const noexcept
+    std::uint32_t model_mesh_part::start_index() const noexcept
     {
         return _start_index;
     }
 
-    std::size_t model_mesh_part::vertex_offset() const noexcept
+    std::uint32_t model_mesh_part::vertex_offset() const noexcept
     {
         return _vertex_offset;
     }
 
-    std::size_t model_mesh_part::vertex_count() const noexcept
+    std::uint32_t model_mesh_part::vertex_count() const noexcept
     {
         return _vertex_count;
     }
 
-    std::size_t model_mesh_part::primitive_count() const noexcept
+    std::uint32_t model_mesh_part::primitive_count() const noexcept
     {
         return _primitive_count;
     }
