+++ source/scener/graphics/model_mesh_part.hpp
@@ -37,20 +37,20 @@ namespace scener::graphics
 
         /// Gets the location in the index array at which to start reading vertices.
         /// \returns location in the index array at which to start reading vertices.
-        std::size_t start_index() const noexcept;
+        std::uint32_t start_index() const noexcept;
 
         /// Gets the offset (in vertices) from the top of vertex buffer.
         /// \returns the offset (in vertices) from the top of vertex buffer.
-        std::size_t vertex_offset() const noexcept;
+        std::uint32_t vertex_offset() const noexcept;
 
         /// Gets the number of vertices used during a draw call.
         /// \returns the number of vertices used during a draw call.
-        std::size_t vertex_count() const noexcept;
+        std::uint32_t vertex_count() const noexcept;
 
         /// Gets the number of primitives to render.
         /// \returns the number of primitives to render. The number of vertices used is a function of primitiveCount and
         ///          primitiveType.
-        std::size_t primitive_count() const noexcept;
+        std::uint32_t primitive_count() const noexcept;
 
         /// Gets the type of primitives to render.
         /// \returns the type of primitives to render.
@@ -60,10 +60,10 @@ namespace scener::graphics
         std::unique_ptr<graphics::index_buffer>     _index_buffer    = { nullptr };
         std::unique_ptr<graphics::vertex_buffer>    _vertex_buffer   = { nullptr };
         std::shared_ptr<graphics::effect_technique> _effect          = { nullptr };
-        std::size_t                                 _start_index     = { 0 };
-        std::size_t                                 _vertex_offset   = { 0 };
-        std::size_t                                 _vertex_count    = { 0 };
-        std::size_t                                 _primitive_count = { 0 };
+        std::uint32_t                               _start_index     = { 0 };
+        std::uint32_t                               _vertex_offset   = { 0 };
+        std::uint32_t                               _vertex_count    = { 0 };
+        std::uint32_t                               _primitive_count = { 0 };
         graphics::primitive_type                    _primitive_type  = { primitive_type::triangle_list };
 
         template <typename T> friend class scener::content::readers::content_type_reader;
